Make a logging conversion explicit
git-svn-id: https://precipitate.googlecode.com/svn/trunk@78 c4e13eb7-e550-0410-89ee-d7df1d48aa01

+++ Shared/GPKeychainItem.m
@@ -109,7 +109,7 @@
   mPassword = nil;
 
   if (result != noErr) {
-    NSLog(@"Couldn't load keychain data (error %d)", result);
+    NSLog(@"Couldn't load keychain data (error %d)", (int)result);
     mUsername = [[NSString alloc] init];
     mService = [[NSString alloc] init];
     mPassword = [[NSString alloc] init];
